feat: find brand by category slug

+++ src/api/upload/upload.service.js
@@ -1,7 +1,10 @@
 const fs = require('fs')
 const cloudinary = require('@configs/cloudinary')
+const { ErrorRes } = require('@/core/error.res')
+const { StatusCodes } = require('http-status-codes')
 const uploadFile = async (file,folder) => {
   try {
+    if (!file) throw new ErrorRes('Not found file', StatusCodes.NOT_FOUND)
     const result = await cloudinary.uploader.upload(file.path,{folder: 'shop_tech'+folder})
     //delete file
     fs.unlinkSync(file.path)
@@ -10,7 +13,7 @@ const uploadFile = async (file,folder) => {
       url: result.secure_url
     }
   } catch (error) {
-    console.log(error);
+    if(error.status === StatusCodes.NOT_FOUND) throw error
     throw new Error('Upload failed')
   }
 }
@@ -18,10 +21,12 @@ const uploadFile = async (file,folder) => {
 const uploadFiles = async (files,folder) => {
   try {
     // const imagesToUp = files.map(file => limit(() => uploadFile(file,folder)))
+    if (files.length === 0) throw new ErrorRes('Not found files', StatusCodes.NOT_FOUND)
     const imagesToUp = files.map(file => uploadFile(file,folder))
     const result = await Promise.all(imagesToUp)
     return result
   } catch (error) {
+    if(error.status === StatusCodes.NOT_FOUND) throw error
     throw new Error('Upload failed')
   }
 }
